pugbot/requirements.txt
discord.py
pendulum
asyncio
aiohttp
